refactor(proxy): dedupe run_* background-task wiring into startup helpers
run_supervisor, run_worker and run_single_process duplicated the
background-task clusters that have historically drifted between
modes; the v1.5.2 worker-mode cert-hotswap bug existed exactly
because one branch was missed (audit H-9, backlog #8).

New binary-local lorica/src/startup/mod.rs owns one source of truth
per cluster:
- start_alerting_stack: notify dispatcher + persisted alert
  dispatcher + probe scheduler + SLA collector/flush (supervisor +
  single-process)
- run_api_server: the common API tail incl. ACME renewal and
  cert-expiry spawns + start_server, so a future mode cannot forget
  to wire renewals (AppState construction stays mode-specific)
- spawn_retention_loop: log/WAF/probe retention + SLA purge
- spawn_health_check_loop: divergences are explicit params
  (backend_connections, config_reload_tx)
- init_geo_resolver_handles: GeoIP/ASN (+ optional rDNS) handle
  registration shared by all three modes

build_notify_dispatcher, spawn_persisted_alert_dispatcher and
run_sla_purge move along (verbatim). A single start(role) mega
helper was deliberately NOT built: the clusters are shared pairwise,
not globally, and a role-switch over ~15 Option parameters would
recreate the AppState nullable-combinatorics the v1.5.9 architecture
audit flags (#42b).

Behaviour-preserving except three documented normalisations of
creation order between independent inits (probe scheduler before
SLA collector, session store before ACME spawns, worker resolver
handles set inside rt.enter()). ~310 duplicated lines eliminated;
main.rs 4693 -> 4380 LOC.
